+++ digits/token_test.go
@@ -1,17 +1,19 @@
 package digits
 
 import (
+	"context"
+	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
+	"strings"
 	"testing"
 
-	"github.com/dghubble/ctxh"
+	"github.com/dghubble/gologin"
 	oauth1Login "github.com/dghubble/gologin/oauth1"
 	"github.com/dghubble/gologin/testutils"
 	"github.com/dghubble/oauth1"
 	"github.com/stretchr/testify/assert"
-	"golang.org/x/net/context"
 )
 
 func TestValidateToken_missingToken(t *testing.T) {
@@ -31,11 +33,13 @@ func TestValidateToken_missingTokenSecret(t *testing.T) {
 func TestTokenHandler(t *testing.T) {
 	proxyClient, _, server := newDigitsTestServer(testAccountJSON)
 	defer server.Close()
+
 	// oauth1 Client will use the proxy client's base Transport
 	ctx := context.WithValue(context.Background(), oauth1.HTTPClient, proxyClient)
 
 	config := &oauth1.Config{}
-	success := func(ctx context.Context, w http.ResponseWriter, req *http.Request) {
+	success := func(w http.ResponseWriter, req *http.Request) {
+		ctx := req.Context()
 		account, err := AccountFromContext(ctx)
 		assert.Nil(t, err)
 		assert.Equal(t, testDigitsToken, account.AccessToken.Token)
@@ -46,35 +50,57 @@ func TestTokenHandler(t *testing.T) {
 		assert.Nil(t, err)
 		assert.Equal(t, testDigitsToken, accessToken)
 		assert.Equal(t, testDigitsSecret, accessSecret)
+		fmt.Fprintf(w, "success handler called")
 	}
-	handler := TokenHandler(config, ctxh.ContextHandlerFunc(success), testutils.AssertFailureNotCalled(t))
-	ts := httptest.NewServer(ctxh.NewHandlerWithContext(ctx, handler))
-	// POST Digits access token to server under test
-	resp, err := http.PostForm(ts.URL, url.Values{accessTokenField: {testDigitsToken}, accessTokenSecretField: {testDigitsSecret}})
-	assert.Nil(t, err)
-	if assert.NotNil(t, resp) {
-		assert.Equal(t, resp.StatusCode, http.StatusOK)
-	}
+
+	// TokenHandler assert that:
+	// - access token/secret are read from POST
+	// - digits account is obtained from the Digits accounts endpoint
+	// - success handler is called
+	// - digits account is added to the success handler ctx
+	tokenHandler := TokenHandler(config, http.HandlerFunc(success), testutils.AssertFailureNotCalled(t))
+	w := httptest.NewRecorder()
+	form := url.Values{accessTokenField: {testDigitsToken}, accessTokenSecretField: {testDigitsSecret}}
+	req, _ := http.NewRequest("POST", "/", strings.NewReader(form.Encode()))
+	req.Header.Add("Content-Type", "application/x-www-form-urlencoded")
+	tokenHandler.ServeHTTP(w, req.WithContext(ctx))
+	assert.Equal(t, http.StatusOK, w.Code)
+	assert.Equal(t, "success handler called", w.Body.String())
 }
 
 func TestTokenHandler_ErrorVerifyingToken(t *testing.T) {
 	proxyClient, server := testutils.NewErrorServer("Digits Account Endpoint Down", http.StatusInternalServerError)
 	defer server.Close()
+
 	// oauth1 Client will use the proxy client's base Transport
 	ctx := context.WithValue(context.Background(), oauth1.HTTPClient, proxyClient)
 
 	config := &oauth1.Config{}
-	handler := TokenHandler(config, testutils.AssertSuccessNotCalled(t), nil)
-	ts := httptest.NewServer(ctxh.NewHandlerWithContext(ctx, handler))
-	// assert that error occurs indicating the Digits Account could not be confirmed
-	resp, _ := http.PostForm(ts.URL, url.Values{accessTokenField: {testDigitsToken}, accessTokenSecretField: {testDigitsSecret}})
-	testutils.AssertBodyString(t, resp.Body, ErrUnableToGetDigitsAccount.Error()+"\n")
+	failure := func(w http.ResponseWriter, req *http.Request) {
+		ctx := req.Context()
+		err := gologin.ErrorFromContext(ctx)
+		if assert.Error(t, err) {
+			assert.Equal(t, err, ErrUnableToGetDigitsAccount)
+		}
+		fmt.Fprintf(w, "failure handler called")
+	}
+
+	// TokenHandler cannot verify Digits account, assert that:
+	// - failure handler is called
+	// - error is added to the failure handler context
+	tokenHandler := TokenHandler(config, testutils.AssertSuccessNotCalled(t), http.HandlerFunc(failure))
+	w := httptest.NewRecorder()
+	form := url.Values{accessTokenField: {testDigitsToken}, accessTokenSecretField: {testDigitsSecret}}
+	req, _ := http.NewRequest("POST", "/", strings.NewReader(form.Encode()))
+	req.Header.Add("Content-Type", "application/x-www-form-urlencoded")
+	tokenHandler.ServeHTTP(w, req.WithContext(ctx))
+	assert.Equal(t, "failure handler called", w.Body.String())
 }
 
 func TestTokenHandler_NonPost(t *testing.T) {
 	config := &oauth1.Config{}
 	handler := TokenHandler(config, testutils.AssertSuccessNotCalled(t), nil)
-	ts := httptest.NewServer(ctxh.NewHandler(handler))
+	ts := httptest.NewServer(handler)
 	resp, err := http.Get(ts.URL)
 	assert.Nil(t, err)
 	// assert that default (nil) failure handler returns a 405 Method Not Allowed
@@ -87,7 +113,7 @@ func TestTokenHandler_NonPost(t *testing.T) {
 func TestTokenHandler_InvalidFields(t *testing.T) {
 	config := &oauth1.Config{}
 	handler := TokenHandler(config, testutils.AssertSuccessNotCalled(t), nil)
-	ts := httptest.NewServer(ctxh.NewHandler(handler))
+	ts := httptest.NewServer(handler)
 
 	// asert errors occur for different missing POST fields
 	resp, err := http.PostForm(ts.URL, url.Values{"wrongFieldName": {testDigitsToken}, accessTokenSecretField: {testDigitsSecret}})