+++ github/login_test.go
@@ -1,18 +1,17 @@
 package github
 
 import (
+	"context"
 	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 
-	"github.com/dghubble/ctxh"
 	"github.com/dghubble/gologin"
 	oauth2Login "github.com/dghubble/gologin/oauth2"
 	"github.com/dghubble/gologin/testutils"
 	"github.com/google/go-github/github"
 	"github.com/stretchr/testify/assert"
-	"golang.org/x/net/context"
 	"golang.org/x/oauth2"
 )
 
@@ -21,13 +20,15 @@ func TestGithubHandler(t *testing.T) {
 	expectedUser := &github.User{ID: github.Int(917408), Name: github.String("Alyssa Hacker")}
 	proxyClient, server := newGithubTestServer(jsonData)
 	defer server.Close()
+
 	// oauth2 Client will use the proxy client's base Transport
 	ctx := context.WithValue(context.Background(), oauth2.HTTPClient, proxyClient)
 	anyToken := &oauth2.Token{AccessToken: "any-token"}
 	ctx = oauth2Login.WithToken(ctx, anyToken)
 
 	config := &oauth2.Config{}
-	success := func(ctx context.Context, w http.ResponseWriter, req *http.Request) {
+	success := func(w http.ResponseWriter, req *http.Request) {
+		ctx := req.Context()
 		githubUser, err := UserFromContext(ctx)
 		assert.Nil(t, err)
 		assert.Equal(t, expectedUser, githubUser)
@@ -40,17 +41,18 @@ func TestGithubHandler(t *testing.T) {
 	// - github User is obtained from the Github API
 	// - success handler is called
 	// - github User is added to the ctx of the success handler
-	githubHandler := githubHandler(config, ctxh.ContextHandlerFunc(success), failure)
+	githubHandler := githubHandler(config, http.HandlerFunc(success), failure)
 	w := httptest.NewRecorder()
 	req, _ := http.NewRequest("GET", "/", nil)
-	githubHandler.ServeHTTP(ctx, w, req)
+	githubHandler.ServeHTTP(w, req.WithContext(ctx))
 	assert.Equal(t, "success handler called", w.Body.String())
 }
 
 func TestGithubHandler_MissingCtxToken(t *testing.T) {
 	config := &oauth2.Config{}
 	success := testutils.AssertSuccessNotCalled(t)
-	failure := func(ctx context.Context, w http.ResponseWriter, req *http.Request) {
+	failure := func(w http.ResponseWriter, req *http.Request) {
+		ctx := req.Context()
 		err := gologin.ErrorFromContext(ctx)
 		if assert.NotNil(t, err) {
 			assert.Equal(t, "oauth2: Context missing Token", err.Error())
@@ -61,10 +63,10 @@ func TestGithubHandler_MissingCtxToken(t *testing.T) {
 	// GithubHandler called without Token in ctx, assert that:
 	// - failure handler is called
 	// - error about ctx missing token is added to the failure handler ctx
-	githubHandler := githubHandler(config, success, ctxh.ContextHandlerFunc(failure))
+	githubHandler := githubHandler(config, success, http.HandlerFunc(failure))
 	w := httptest.NewRecorder()
 	req, _ := http.NewRequest("GET", "/", nil)
-	githubHandler.ServeHTTP(context.Background(), w, req)
+	githubHandler.ServeHTTP(w, req)
 	assert.Equal(t, "failure handler called", w.Body.String())
 }
 
@@ -78,7 +80,8 @@ func TestGithubHandler_ErrorGettingUser(t *testing.T) {
 
 	config := &oauth2.Config{}
 	success := testutils.AssertSuccessNotCalled(t)
-	failure := func(ctx context.Context, w http.ResponseWriter, req *http.Request) {
+	failure := func(w http.ResponseWriter, req *http.Request) {
+		ctx := req.Context()
 		err := gologin.ErrorFromContext(ctx)
 		if assert.NotNil(t, err) {
 			assert.Equal(t, ErrUnableToGetGithubUser, err)
@@ -89,10 +92,10 @@ func TestGithubHandler_ErrorGettingUser(t *testing.T) {
 	// GithubHandler cannot get Github User, assert that:
 	// - failure handler is called
 	// - error cannot get Github User added to the failure handler ctx
-	githubHandler := githubHandler(config, success, ctxh.ContextHandlerFunc(failure))
+	githubHandler := githubHandler(config, success, http.HandlerFunc(failure))
 	w := httptest.NewRecorder()
 	req, _ := http.NewRequest("GET", "/", nil)
-	githubHandler.ServeHTTP(ctx, w, req)
+	githubHandler.ServeHTTP(w, req.WithContext(ctx))
 	assert.Equal(t, "failure handler called", w.Body.String())
 }